feat(ui): Retool our UI stuff
