+++ addons/ui/RscDisplayClient.inc
@@ -1,0 +1,12 @@
+class RscDisplayClient : RscStandardDisplay {
+	class ControlsBackground {
+		class LoadingPic : RscPicture {
+			idc = -1;
+			x = "SafeZoneX";
+			y = "SafeZoneY";
+			h = "SafeZoneH";
+			w = "SafeZoneW";
+			text = QPATHTOF(data\titlescreen-drop.jpg);
+		};
+	};
+};
