+++ addons/ui/RscDisplayClientWait.inc
@@ -1,0 +1,10 @@
+//--- I believe this is while connecting to a multiplayer mission.
+class RscDisplayClientWait : RscDisplayMPPlayers {
+  class LoadingPic : RscPicture {
+    x = "SafeZoneX";
+    y = "SafeZoneY";
+    h = "SafeZoneH";
+    w = "SafeZoneW";
+    text = QPATHTOF(data\titlescreen-drop.jpg);
+  };
+};
