+++ addons/ui/RscDisplayLoading.inc
@@ -1,0 +1,18 @@
+//--- Sampled for all loading menus.
+class RscDisplayLoading {
+	class Variants {
+		class SWS_Loading : RscStandardDisplay {
+      idd = 250;
+			class ControlsBackground {
+				class LoadingPic : RscPicture {
+					idc = -1;
+					x = "SafeZoneX";
+					y = "SafeZoneY";
+					h = "SafeZoneH";
+					w = "SafeZoneW";
+					text = QPATHTOF(data\titlescreen-drop.jpg);
+				};
+			};
+		};
+	};
+};
