+++ addons/ui/config.cpp
@@ -1,20 +1,35 @@
 #include "script_component.hpp"
 
-#include "RscDisplayMain.hpp"
-#include "CfgLoadingScreens.hpp"
+class Controls;
+class RscActivePicture;
+class RscActiveText;
+class RscControlsGroup;
+class RscControlsGroupNoScrollbars;
+class RscDisplayMPPlayers;
+class RscMainMenuSpotlight;
+class RscPicture;
+class RscPictureKeepAspect;
+class RscStandardDisplay;
+class RscText;
 
-class CfgPatches
-{
-    class ADDON
-    {
-        name = QUOTE(COMPONENT);
-        units[] = { };
-        weapons[] = { };
-        magazines[] = { };
-        ammo[] = { };
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = { "19th_Fleet_Armor" };
-        author = QUOTE(AUTHOR);
-        VERSION_CONFIG;
-    };
+#include "RscDisplayClient.inc"
+#include "RscDisplayClientWait.inc"
+#include "RscDisplayLoadMission.inc"
+#include "RscDisplayMain.inc"
+#include "RscDisplayStart.inc"
+
+#include "RscDisplayLoading.inc"
+
+class CfgPatches {
+  class ADDON {
+    name = QUOTE(COMPONENT);
+    units[] = { };
+    weapons[] = { };
+    magazines[] = { };
+    ammo[] = { };
+    requiredVersion = REQUIRED_VERSION;
+    requiredAddons[] = { "19th_Fleet_Armor" };
+    author = QUOTE(AUTHOR);
+    VERSION_CONFIG;
+  };
 };
